# tests/yoke/http/golden/openapi.json
{
  "components": {
    "schemas": {
      "ActiveRuntimeInfo": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "state": {
            "enum": [
              "idle",
              "running",
              "stopping",
              "waiting_input",
              "error"
            ],
            "title": "State",
            "type": "string"
          },
          "turnID": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnid"
          }
        },
        "required": [
          "state"
        ],
        "title": "ActiveRuntimeInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterruptData": {
        "additionalProperties": false,
        "properties": {
          "interrupted": {
            "title": "Interrupted",
            "type": "boolean"
          },
          "turnID": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnid"
          }
        },
        "required": [
          "interrupted"
        ],
        "title": "InterruptData",
        "type": "object"
      },
      "InterruptResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InterruptData"
          }
        },
        "required": [
          "data"
        ],
        "title": "InterruptResponse",
        "type": "object"
      },
...
      "PromptAdmissionReceipt": {
        "additionalProperties": false,
        "properties": {
          "admittedSeq": {
            "title": "Admittedseq",
            "type": "integer"
          },
          "delivery": {
            "enum": [
              "steer",
              "queue"
            ],
            "title": "Delivery",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "promotedSeq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promotedseq"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptInput"
          },
          "sessionID": {
            "title": "Sessionid",
            "type": "string"
          },
          "state": {
            "enum": [
              "admitted",
              "promoted",
              "removed"
            ],
            "title": "State",
            "type": "string"
          },
          "timeCreated": {
            "title": "Timecreated",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionID",
          "prompt",
          "delivery",
          "state",
          "admittedSeq",
          "timeCreated"
        ],
        "title": "PromptAdmissionReceipt",
        "type": "object"
      },
      "PromptAdmissionRequest": {
        "additionalProperties": false,
        "properties": {
          "delivery": {
            "default": "steer",
            "enum": [
              "steer",
              "queue"
            ],
            "title": "Delivery",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptInput"
          },
          "resume": {
            "default": true,
            "title": "Resume",
            "type": "boolean"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "PromptAdmissionRequest",
        "type": "object"
      },
      "PromptAdmissionResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PromptAdmissionReceipt"
          }
        },
        "required": [
          "data"
        ],
        "title": "PromptAdmissionResponse",
        "type": "object"
      },
      "PromptAttachment": {
        "additionalProperties": false,
        "properties": {
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "const": "file",
            "default": "file",
            "title": "Type",
            "type": "string"
          },
          "uri": {
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "uri",
          "name",
          "mime"
        ],
        "title": "PromptAttachment",
        "type": "object"
      },
      "PromptInput": {
        "additionalProperties": false,
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/PromptAttachment"
            },
            "title": "Attachments",
            "type": "array"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          }
        },
        "title": "PromptInput",
        "type": "object"
      },
...
      "QueueData": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "items"
        ],
        "title": "QueueData",
        "type": "object"
      },
      "QueueDeliveryOperation": {
        "additionalProperties": false,
        "properties": {
          "delivery": {
            "enum": [
              "steer",
              "queue"
            ],
            "title": "Delivery",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "op": {
            "const": "setDelivery",
            "title": "Op",
            "type": "string"
          }
        },
        "required": [
          "op",
          "id",
          "delivery"
        ],
        "title": "QueueDeliveryOperation",
        "type": "object"
      },
      "QueueItem": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "delivery": {
            "enum": [
              "steer",
              "queue"
            ],
            "title": "Delivery",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptInput"
          },
          "state": {
            "const": "admitted",
            "default": "admitted",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "prompt",
          "delivery",
          "paused",
          "createdAt"
        ],
        "title": "QueueItem",
        "type": "object"
      },
      "QueueMoveAfterOperation": {
        "additionalProperties": false,
        "properties": {
          "afterID": {
            "title": "Afterid",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "op": {
            "const": "moveAfter",
            "title": "Op",
            "type": "string"
          }
        },
        "required": [
          "op",
          "id",
          "afterID"
        ],
        "title": "QueueMoveAfterOperation",
        "type": "object"
      },
      "QueueMoveBeforeOperation": {
        "additionalProperties": false,
        "properties": {
          "beforeID": {
            "title": "Beforeid",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "op": {
            "const": "moveBefore",
            "title": "Op",
            "type": "string"
          }
        },
        "required": [
          "op",
          "id",
          "beforeID"
        ],
        "title": "QueueMoveBeforeOperation",
        "type": "object"
      },
      "QueueMoveToStartOperation": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "op": {
            "const": "moveToStart",
            "title": "Op",
            "type": "string"
          }
        },
        "required": [
          "op",
          "id"
        ],
        "title": "QueueMoveToStartOperation",
        "type": "object"
      },
      "QueueOperation": {
        "discriminator": {
          "mapping": {
            "moveAfter": "#/components/schemas/QueueMoveAfterOperation",
            "moveBefore": "#/components/schemas/QueueMoveBeforeOperation",
            "moveToStart": "#/components/schemas/QueueMoveToStartOperation",
            "remove": "#/components/schemas/QueueRemoveOperation",
            "setDelivery": "#/components/schemas/QueueDeliveryOperation",
            "setPaused": "#/components/schemas/QueuePausedOperation",
            "update": "#/components/schemas/QueueUpdateOperation"
          },
          "propertyName": "op"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/QueueUpdateOperation"
          },
          {
            "$ref": "#/components/schemas/QueueDeliveryOperation"
          },
          {
            "$ref": "#/components/schemas/QueuePausedOperation"
          },
          {
            "$ref": "#/components/schemas/QueueRemoveOperation"
          },
          {
            "$ref": "#/components/schemas/QueueMoveBeforeOperation"
          },
          {
            "$ref": "#/components/schemas/QueueMoveAfterOperation"
          },
          {
            "$ref": "#/components/schemas/QueueMoveToStartOperation"
          }
        ]
      },
      "QueuePatchRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedRevision": {
            "minimum": 0.0,
            "title": "Expectedrevision",
            "type": "integer"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/QueueOperation"
            },
            "title": "Operations",
            "type": "array"
          }
        },
        "required": [
          "expectedRevision",
          "operations"
        ],
        "title": "QueuePatchRequest",
        "type": "object"
      },
      "QueuePausedOperation": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "op": {
            "const": "setPaused",
            "title": "Op",
            "type": "string"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          }
        },
        "required": [
          "op",
          "id",
          "paused"
        ],
        "title": "QueuePausedOperation",
        "type": "object"
      },
      "QueueRemoveOperation": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "op": {
            "const": "remove",
            "title": "Op",
            "type": "string"
          }
        },
        "required": [
          "op",
          "id"
        ],
        "title": "QueueRemoveOperation",
        "type": "object"
      },
      "QueueResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QueueData"
          }
        },
        "required": [
          "data"
        ],
        "title": "QueueResponse",
        "type": "object"
      },
      "QueueUpdateOperation": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "op": {
            "const": "update",
            "title": "Op",
            "type": "string"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptInput"
          }
        },
        "required": [
          "op",
          "id",
          "prompt"
        ],
        "title": "QueueUpdateOperation",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WaitResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ActiveRuntimeInfo"
          }
        },
        "required": [
          "data"
        ],
        "title": "WaitResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Yoke HTTP API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session/{session_id}/interrupt": {
      "post": {
        "operationId": "interruptSession",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterruptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Interrupt Session"
      }
    },
...
    "/api/v1/session/{session_id}/prompt": {
      "post": {
        "operationId": "admitSessionPrompt",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptAdmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptAdmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admit Prompt"
      }
    },
...
    "/api/v1/session/{session_id}/queue": {
      "get": {
        "operationId": "getSessionQueue",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Queue"
      },
      "patch": {
        "operationId": "patchSessionQueue",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueuePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Queue"
      }
    },
...
    "/api/v1/session/{session_id}/wait": {
      "post": {
        "operationId": "waitForSession",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timeoutMs",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 300000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timeoutms"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wait For Session"
      }
    },
...
  }
}